Refactor options window and persist settings
